Moving phone request right after phone button click
